+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @serwist/svelte
 
+## 9.5.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @serwist/build@9.5.10
+  - @serwist/utils@9.5.10
+  - @serwist/window@9.5.10
+  - serwist@9.5.10
+
 ## 9.5.9
 
 ### Patch Changes