+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @serwist/vite
 
+## 9.5.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @serwist/build@9.5.10
+  - @serwist/utils@9.5.10
+  - serwist@9.5.10
+
 ## 9.5.9
 
 ### Patch Changes